+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/auth/AuthUser.kt
@@ -2,6 +2,5 @@ package com.teamsparta.exhibitionnewsfeed.domain.auth
 
 data class AuthUser(
     val id: Long,
-    val token: String,
-    val tokenType: TokenType
+    val token: String
 )
