+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/auth/service/AuthService.kt
@@ -9,4 +9,5 @@ interface AuthService {
     fun login(request: LoginRequest): LoginResponse
     fun getNewAccessToken(authUser: AuthUser): String
     fun logout(authUser: AuthUser)
+    fun verifyPassword(userId: Long, password: String?): Boolean
 }
