+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/likes/controller/CommentLikeController.kt
@@ -2,7 +2,6 @@ package com.teamsparta.exhibitionnewsfeed.domain.likes.controller
 
 import com.teamsparta.exhibitionnewsfeed.domain.auth.AuthUser
 import com.teamsparta.exhibitionnewsfeed.domain.auth.RequestUser
-import com.teamsparta.exhibitionnewsfeed.domain.auth.TokenType
 import com.teamsparta.exhibitionnewsfeed.domain.likes.dto.CommentLikeRequest
 import com.teamsparta.exhibitionnewsfeed.domain.likes.service.LikeService
 import com.teamsparta.exhibitionnewsfeed.exception.UnauthorizedException
@@ -25,7 +24,6 @@ class CommentLikeController(
         @RequestUser @Parameter(hidden = true) authUser: AuthUser,
         @RequestBody commentLikeRequest: CommentLikeRequest
     ): ResponseEntity<Unit> {
-        checkAccessToken(authUser)
         if (authUser.id != commentLikeRequest.userId)
             throw UnauthorizedException("You are not allowed to use this comment like")
         likeService.likeComment(postId, commentId, authUser, commentLikeRequest.userId)
@@ -40,15 +38,10 @@ class CommentLikeController(
         @RequestUser @Parameter(hidden = true) authUser: AuthUser,
         @RequestBody commentLikeRequest: CommentLikeRequest
     ): ResponseEntity<Unit> {
-        checkAccessToken(authUser)
         if (authUser.id != commentLikeRequest.userId)
             throw UnauthorizedException("You are not allowed to use this comment like")
         likeService.removeCommentLike(postId, commentId, likeId, commentLikeRequest.userId)
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build()
-    }
-
-    private fun checkAccessToken(authUser: AuthUser) {
-        if (authUser.tokenType != TokenType.ACCESS_TOKEN) throw UnauthorizedException("유효한 토큰이 아닙니다.")
     }
 }
