+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/likes/controller/PostLikeController.kt
@@ -2,10 +2,8 @@ package com.teamsparta.exhibitionnewsfeed.domain.likes.controller
 
 import com.teamsparta.exhibitionnewsfeed.domain.auth.AuthUser
 import com.teamsparta.exhibitionnewsfeed.domain.auth.RequestUser
-import com.teamsparta.exhibitionnewsfeed.domain.auth.TokenType
 import com.teamsparta.exhibitionnewsfeed.domain.likes.dto.PostLikeResponse
 import com.teamsparta.exhibitionnewsfeed.domain.likes.service.LikeService
-import com.teamsparta.exhibitionnewsfeed.exception.UnauthorizedException
 import io.swagger.v3.oas.annotations.Parameter
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -22,7 +20,6 @@ class PostLikeController(
         @PathVariable postId: Long,
         @RequestUser @Parameter(hidden = true) authUser: AuthUser,
     ): ResponseEntity<Unit> {
-        checkAuth(authUser)
         likeService.likePost(postId, authUser)
 
         return ResponseEntity
@@ -35,9 +32,7 @@ class PostLikeController(
         @PathVariable likeId: Long,
         @PathVariable postId: Long,
         @RequestUser @Parameter(hidden = true) authUser: AuthUser
-    ):
-            ResponseEntity<Unit> {
-        checkAuth(authUser)
+    ): ResponseEntity<Unit> {
         likeService.removePostLike(likeId, postId)
 
         return ResponseEntity
@@ -48,9 +43,5 @@ class PostLikeController(
     @GetMapping("/posts/{postId}")
     fun getLikes(@PathVariable postId: Long): List<PostLikeResponse> {
         return likeService.getLikesByPostId(postId)
-    }
-
-    private fun checkAuth(authUser: AuthUser) {
-        if (authUser.tokenType != TokenType.ACCESS_TOKEN) throw UnauthorizedException("유효한 토큰이 아닙니다.")
     }
 }
