+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/newsfeed/comment/controller/CommentController.kt
@@ -2,12 +2,10 @@ package com.teamsparta.exhibitionnewsfeed.domain.newsfeed.comment.controller
 
 import com.teamsparta.exhibitionnewsfeed.domain.auth.AuthUser
 import com.teamsparta.exhibitionnewsfeed.domain.auth.RequestUser
-import com.teamsparta.exhibitionnewsfeed.domain.auth.TokenType
 import com.teamsparta.exhibitionnewsfeed.domain.newsfeed.comment.dto.CommentResponse
 import com.teamsparta.exhibitionnewsfeed.domain.newsfeed.comment.dto.CreateCommentRequest
 import com.teamsparta.exhibitionnewsfeed.domain.newsfeed.comment.dto.UpdateCommentRequest
 import com.teamsparta.exhibitionnewsfeed.domain.newsfeed.comment.service.CommentService
-import com.teamsparta.exhibitionnewsfeed.exception.UnauthorizedException
 import io.swagger.v3.oas.annotations.Parameter
 import jakarta.validation.Valid
 import org.springframework.http.HttpStatus
@@ -25,7 +23,6 @@ class CommentController(
         @RequestUser @Parameter(hidden = true) authUser: AuthUser,
         @RequestBody request: CreateCommentRequest
     ): ResponseEntity<CommentResponse> {
-        checkAccessToken(authUser)
         return ResponseEntity
             .status(HttpStatus.CREATED)
             .body(commentService.createComment(postId, authUser, request))
@@ -38,7 +35,6 @@ class CommentController(
         @RequestUser @Parameter(hidden = true) authUser: AuthUser,
         @RequestBody @Valid request: UpdateCommentRequest
     ): ResponseEntity<CommentResponse> {
-        checkAccessToken(authUser)
         return ResponseEntity
             .status(HttpStatus.OK)
             .body(commentService.updateComment(postId, commentId, authUser, request))
@@ -50,14 +46,9 @@ class CommentController(
         @RequestUser @Parameter(hidden = true) authUser: AuthUser,
         @PathVariable commentId: Long
     ): ResponseEntity<Unit> {
-        checkAccessToken(authUser)
         commentService.deleteComment(postId, commentId, authUser)
         return ResponseEntity
             .status(HttpStatus.NO_CONTENT)
             .build()
-    }
-
-    private fun checkAccessToken(authUser: AuthUser) {
-        if (authUser.tokenType != TokenType.ACCESS_TOKEN) throw UnauthorizedException("유효한 토큰이 아닙니다.")
     }
 }
