+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/user/controller/UserController.kt
@@ -2,7 +2,6 @@ package com.teamsparta.exhibitionnewsfeed.domain.user.controller
 
 import com.teamsparta.exhibitionnewsfeed.domain.auth.AuthUser
 import com.teamsparta.exhibitionnewsfeed.domain.auth.RequestUser
-import com.teamsparta.exhibitionnewsfeed.domain.auth.TokenType
 import com.teamsparta.exhibitionnewsfeed.domain.user.dto.SignUpRequest
 import com.teamsparta.exhibitionnewsfeed.domain.user.dto.SignUpResponse
 import com.teamsparta.exhibitionnewsfeed.domain.user.dto.UpdateUserProfileRequest
@@ -20,7 +19,6 @@ import org.springframework.web.bind.annotation.*
 class UserController(
     private val userService: UserService
 ) {
-
     @PostMapping("/sign-up")
     fun signUp(@RequestBody @Valid request: SignUpRequest): ResponseEntity<SignUpResponse> {
         return ResponseEntity
@@ -35,33 +33,15 @@ class UserController(
             .body(userService.getProfile(userId))
     }
 
-    @PostMapping("/profile/{userId}/verify")
-    fun verifyPassword(
-        @PathVariable userId: Long,
-        @RequestUser @Parameter(hidden = true) authUser: AuthUser,
-        @RequestBody request: Map<String, String>
-    ): ResponseEntity<Any> {
-        checkAuth(userId, authUser)
-        if (!userService.verifyPassword(userId, request["password"])) throw UnauthorizedException("비밀번호가 일치하지 않습니다.")
-
-        //TODO 암호화된 검증 토큰 response에 넣기
-        return ResponseEntity.status(HttpStatus.OK).build()
-    }
-
     @PatchMapping("/profile/{userId}")
     fun updateProfile(
         @PathVariable userId: Long,
         @RequestUser @Parameter(hidden = true) authUser: AuthUser,
         @RequestBody request: UpdateUserProfileRequest
     ): ResponseEntity<UserProfileResponse> {
-        checkAuth(userId, authUser)
+        if (userId != authUser.id) throw UnauthorizedException("권한이 없습니다.")
         return ResponseEntity
             .status(HttpStatus.OK)
             .body(userService.updateProfile(userId, request))
-    }
-
-    private fun checkAuth(userId: Long, authUser: AuthUser) {
-        if (userId != authUser.id) throw UnauthorizedException("권한이 없습니다.")
-        if (authUser.tokenType != TokenType.ACCESS_TOKEN) throw UnauthorizedException("유효한 토큰이 아닙니다.")
     }
 }
