+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/user/service/UserServiceImpl.kt
@@ -29,12 +29,6 @@ class UserServiceImpl(
         return UserProfileResponse.from(user)
     }
 
-    override fun verifyPassword(userId: Long, password: String?): Boolean {
-        val user =
-            userRepository.findByIdOrNull(userId) ?: throw ModelNotFoundException("User id $userId not found.", userId)
-        return user.isValidPassword(password ?: "", passwordEncoder)
-    }
-
     @Transactional
     override fun updateProfile(userId: Long, request: UpdateUserProfileRequest): UserProfileResponse {
         val user =
